// ui-app/src/app/api/models/school-details-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface SchoolDetailsResponseDto {
  enrollmentId?: number;
  enrollmentYear?: number;
  examCentreId?: number;
  examCentreName?: string;
  id?: number;
  name?: string;
  vigyanKendraId?: number;
  vigyanKendraName?: string;
}