// ui-app/src/app/api/models/student-class-details-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { SubjectDetailsResponseDto } from '../models/subject-details-response-dto';
export interface StudentClassDetailsResponseDto {
  id?: number;
  name?: string;
  subjects?: Array<SubjectDetailsResponseDto>;
}