// ui-app/src/app/api/models/vigyan-kendra-details.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ExaminationCentreDetails } from '../models/examination-centre-details';
import { SchoolDetails } from '../models/school-details';
import { Student } from '../models/student';
import { UserDetails } from '../models/user-details';
export interface VigyanKendraDetails {
  code?: string;
  createdBy?: string;
  createdDate?: string;
  examinationCentres?: Array<ExaminationCentreDetails>;
  id?: number;
  modifiedBy?: string;
  modifiedDate?: string;
  name?: string;
  schoolDetails?: Array<SchoolDetails>;
  students?: Array<Student>;
  users?: Array<UserDetails>;
}